fix travis.yml
