+++ app.py
@@ -1,4 +1,5 @@
 import os
+from site_file_validator import SitefileValidator
 
 from flask import Flask
 
@@ -11,6 +12,9 @@ if os.path.exists(os.path.join(PROJECT_DIR, '.env')):
     application.config.from_pyfile('.env')
 
 from services import *
+
+v = SitefileValidator()
+v.allow_unknown = True
 
 if __name__ == '__main__':
     application.run()
